[NFC reformat] Applies pre-commit formatting to Python files.
This is a large change because prior to this point, Python files in the project were not consistently formatted. This reformats them all with black defaults.

Based on experience with prior projects, if you have a dev/long-term branch with Python patches, you can minimize merge conflicts prior to rebasing to include this commit by running `black` on your modified Python files, squashing, and then rebasing/merging.