[torch-mlir][sparse] replace ad-hoc mechanism with proper FX export
Now that the PyDev feature request pytorch/pytorch#117188 has
been completed, we can remove all the ad-hoc code that propagates
sparsity metadata and replace it with the built-int PyDev metadata
for sparse tensors. This removes a lot of code and also ensures
sparsity is consistent with the torch.sparse package for all cases.