Include `torch-mlir-opt` in Python wheels (#3964)
This adds the `torch-mlir-opt` tool to the Python wheels, which allows
to use the commandline tool via a pip installed package instead of
having to compile the torch-mlir project yourself. The executable is
still installed to the deault location and copied over via the
`setup.py` to be included in the Python wheel. This could be refactored
and handled within CMake in a follow-up.