[ONNX][TORCH] Add Onnx->Linalg lowering for RotaryEmbedding Op (#4002)
This commit adds the Onnx->Linalg lowering for Onnx's RotaryEmbedding op
(ref:
https://github.com/microsoft/onnxruntime/blob/main/docs/ContribOperators.md#commicrosoftrotaryembedding)
by registering a customized torch op named
`OnnxVariantAtenRotaryEmbeddingOp`. This is done so that the Onnx's
RotaryEmbedding op can be lowered to this op and this op can be lowered
from Torch->Linalg.

The lowering has been adopted from the OnnxRuntime. Files for
references:
1.)
https://github.com/microsoft/onnxruntime/blob/e1e3f623f61816008e79dddc91a51ffe7f0ff5cf/onnxruntime/contrib_ops/cpu/bert/rotary_embedding.cc#L47-L93
2.)
https://github.com/microsoft/onnxruntime/blob/94c69f55d480cb4a8dcbc161d29ef3acca9392a7/onnxruntime/contrib_ops/cpu/bert/rotary_embedding_helper.h

---------

Signed-off-by: Vivek Khandelwal <vivekkhandelwal1424@gmail.com>
Co-authored-by: zjgarvey <47986913+zjgarvey@users.noreply.github.com>