Extend to 1D and 3D the torch to linalg lowering of the average pool operator with count_include_pad = false (#4035)
Currently the avg_pool2d PyTorch operation supports the cases where
count_include_pad is true and false, but the avg_pool1d and avg_pool3d
only the true case is supported (which is simpler).

The count_include_pad = false support for avg_pool2d was added by
@AmosLewis in this change (reviewed by @rsuderman and @nirvedhmeshram) :
https://github.com/llvm/torch-mlir/pull/3235

In this change I generalized the logic added above. I also did some
refactoring to the original code to reduce the size of the functions and
to avoid redundancy when possible.

@sahas3 @dixinzhou @rafaelubal

---------

Co-authored-by: Ivan Garcia <igarcia@vdi-ah2ddp-178.dhcp.mathworks.com>