Add Windows build workflow in the CI. (#4041)
1. Adds a workflow to trigger windows build in the CI similar to the
existing one for Linux to address the build CI request captured in
https://github.com/llvm/torch-mlir/issues/3985
2. Only non-python based LIT unit tests are enabled in the workflow due
to failures reported in https://github.com/llvm/torch-mlir/issues/4040
3. The CI is only run against `torch-nightly` since python based tests
are disabled. My understanding is that without the python based tests
there's no need to run against the `stable` version too. If that's not
the case, will enable `stable` as well. Tested that both `nightly` and
`stable` versions work fine:
https://github.com/sahas3/torch-mlir/actions/runs/13477124004