[TorchToLinalg] simplify non-broadcast unit dim indexing maps in elementise generics (#4107)
This change is made to reduce the pattern-matching load for fusing
elementwise generic ops with non-broadcasting unit dims.

For example, adding tensors with shapes `[6,1]` and `[1]`, the output
shape will be `[6,1]`. Before this change, the indexing maps were
inconsistent between the inputs and outputs for the unit-dim (constant 0
for inputs, and a dim expression for the output).

---------

Signed-off-by: zjgarvey <zjgarvey@gmail.com>