Bring the C++ ONNX importer on par with `onnx_importer.py` (#3960)
This PR heavily refactors the C++ importer to make it a viable (and
faster) alternative to the Python one for those who need to import ONNX
models and want to avoid depending on the python ecosystem.

- The C++ importer now outputs the same exact mlir as the Python one
(tested on `alt_e2eshark` test suite). Achieving perfect output matches
required to introduce an associative map iterable according to insertion
order (to mimic `Dict` in Python).
- The code tries to mirror 1-to-1 the Python counterpart whenever
possible/convenient.
- Adds support for embedding ONNX external data in the mlir. This
functionality is not part of torch-mlir's `onnx_importer.py` but of
IREE's `import_onnx`.
- Efforts have been made to remove the direct dependency on LLVM support
lib. There is however a transitive dependency on such lib through
`MLIRCAPIIR` and `TorchMLIRCAPI` (MLIR libraries uniformly depend on
`LLVMSupport`).