build: manually update PyTorch version (#4112)
This commit sets the PyTorch and TorchVision versions to nightly release
2025-04-23.

This commit also updates the fx_importer tests with the changes related
to symbolic shape ids introduced in
https://github.com/pytorch/pytorch/commit/f649ee73ce28fe8e021ec93f69db4eb8f14285f4.

This commit disables some checks because of the mismatch in the
resultant IR for PyTorch nightly and stable versions. Those checks will
be enabled once they are on the same page.

---------

Signed-off-by: Vivek Khandelwal <vivekkhandelwal1424@gmail.com>